requirements.txt
requests>=2.31.0
pillow>=10.0.0
numpy<2.0.0
torch
pydantic>=2.0.0
google-genai>=1.0.0
